+++ NEWS.md
@@ -1,10 +1,13 @@
 # zarr (development version)
 
 - AWS S3 store access added for reading and, with appropriate authentication, writing. The function `s3_list_dir()` can be used to walk the directory listing of an S3 bucket to locate Zarr stores.
+- New `zarr_array::resize()` method with which arrays can be resized, growing or shrinking across all dimensions simultaneously, including on the lower end of the dimensions. When changing the "low" side of the dimensions, which is always by full chunks, the indexing of all data changes to use the new origin.
+- New `zarr_array::promote()` method which adds a dimension to the array.
 - New `zarr_object` base class for name and attribute management.
 - `dirty` field can be explicitly set on any node to force writing of the metadata to the store.
 - `chunk_key_encoding` is automatically set in array metadata.
 - New `optimal_chunking()` function to determine optimal chunking for an array, optionally applying different weights to dimensions or groups of dimensions.
+- Better support for scalar arrays. Scalar arrays can now be constructed by the `array_builder` class and written to like regular arrays.
 - Testing expanded.
 - Documentation expanded and updated.